fix: refactoring token routes
