+++ applications/app-service/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,5 @@
+{"properties": [{
+  "name": "aws.dynamodb.endpoint",
+  "type": "java.lang.String",
+  "description": "A description for 'aws.dynamodb.endpoint'"
+}]}
